RagService/requirements.txt
python-dotenv~=1.2.1
chromadb~=1.4.0
ollama~=0.6.1